feat: extract CalDAV sync functionality with polling for import status
